+++ requests_cache/backends/filesystem.py
@@ -1,11 +1,12 @@
 from contextlib import contextmanager
+from glob import glob
 from os import listdir, makedirs, unlink
-from os.path import abspath, dirname, expanduser, isabs, join, splitext
+from os.path import abspath, basename, expanduser, isabs, join, splitext
 from pathlib import Path
 from pickle import PickleError
 from shutil import rmtree
 from tempfile import gettempdir
-from typing import Union
+from typing import List, Union
 
 from ..serializers import SERIALIZERS
 from . import BaseCache, BaseStorage
@@ -28,8 +29,14 @@ class FileCache(BaseCache):
     def __init__(self, cache_name: Union[Path, str] = 'http_cache', use_temp: bool = False, **kwargs):
         super().__init__(**kwargs)
         self.responses = FileDict(cache_name, use_temp=use_temp, **kwargs)
-        db_path = join(dirname(self.responses.cache_dir), 'redirects.sqlite')
+        db_path = join(self.responses.cache_dir, 'redirects.sqlite')
         self.redirects = DbDict(db_path, 'redirects', **kwargs)
+
+    def clear(self):
+        """Clear the cache"""
+        # FileDict.clear() removes and re-creates the cache directory, including redirects.sqlite
+        self.responses.clear()
+        self.redirects.init_db()
 
 
 class FileDict(BaseStorage):
@@ -79,8 +86,7 @@ class FileDict(BaseStorage):
                 f.write(self.serializer.dumps(value))
 
     def __iter__(self):
-        for filename in listdir(self.cache_dir):
-            yield splitext(filename)[0]
+        yield from self.keys()
 
     def __len__(self):
         return len(listdir(self.cache_dir))
@@ -90,10 +96,12 @@ class FileDict(BaseStorage):
             rmtree(self.cache_dir, ignore_errors=True)
             makedirs(self.cache_dir)
 
-    def paths(self):
-        """Get file paths to all cached responses"""
-        for key in self:
-            yield self._path(key)
+    def keys(self):
+        return [splitext(basename(path))[0] for path in self.paths()]
+
+    def paths(self) -> List[str]:
+        """Get absolute file paths to all cached responses"""
+        return glob(self._path('*'))
 
 
 def _get_cache_dir(cache_dir: Union[Path, str], use_temp: bool) -> str: